+++ src/potential/EAM/eam.cpp
@@ -1,4 +1,16 @@
+// Copyright © 2020 Conor Williams <conorwilliams@outlook.com>
 
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+// This file is part of openFLY.
+
+// OpenFLY is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License
+// as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+
+// OpenFLY is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+// warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License along with openFLY. If not, see <https://www.gnu.org/licenses/>.
 
 #include "libfly/potential/EAM/eam.hpp"
 
@@ -38,7 +50,9 @@ namespace fly::potential {
     return (0.5 * v_sum) + f_sum;
   }
 
-  void EAM::gradient(system::SoA<PotentialGradient&> out, system::SoA<TypeID const&, Frozen const&> in, neigh::List const& nl,
+  void EAM::gradient(system::SoA<PotentialGradient&> out,
+                     system::SoA<TypeID const&, Frozen const&> in,
+                     neigh::List const& nl,
                      int num_threads) {
     //
     verify(in.size() == out.size(), "EAM gradient size mismatch in={} out={}", in.size(), out.size());